feat: consolidate jdbc benchmark ResultSet collecting
