+++ benchmarks/src/jmh/kotlin/com/github/kdbc/benchmarks/postgresql/JdbcBenchmarkSingle.kt
@@ -1,6 +1,5 @@
 package com.github.kdbc.benchmarks.postgresql
 
-import kotlinx.datetime.toKotlinLocalDateTime
 import org.openjdk.jmh.annotations.Benchmark
 import org.openjdk.jmh.annotations.BenchmarkMode
 import org.openjdk.jmh.annotations.Fork
@@ -13,7 +12,6 @@ import org.openjdk.jmh.annotations.State
 import org.openjdk.jmh.annotations.TearDown
 import org.openjdk.jmh.annotations.Warmup
 import java.sql.Connection
-import java.time.LocalDateTime
 import java.util.concurrent.TimeUnit
 
 @Warmup(iterations = 4, time = 10, timeUnit = TimeUnit.SECONDS)
@@ -40,31 +38,13 @@ open class JdbcBenchmarkSingle {
         if (id > 5000) id = 1
     }
 
-//    @Benchmark
+    @Benchmark
     open fun queryData() {
         step()
         connection.prepareStatement(jdbcQuerySingle).use { preparedStatement ->
             preparedStatement.setInt(1, id)
             preparedStatement.executeQuery().use { resultSet ->
-                val items = mutableListOf<PostDataClass>()
-                while (resultSet.next()) {
-                    val item = PostDataClass(
-                        resultSet.getInt(1),
-                        resultSet.getString(2),
-                        resultSet.getObject(3, LocalDateTime::class.java).toKotlinLocalDateTime(),
-                        resultSet.getObject(4, LocalDateTime::class.java).toKotlinLocalDateTime(),
-                        resultSet.getInt(5),
-                        resultSet.getInt(6),
-                        resultSet.getInt(7),
-                        resultSet.getInt(8),
-                        resultSet.getInt(9),
-                        resultSet.getInt(10),
-                        resultSet.getInt(11),
-                        resultSet.getInt(12),
-                        resultSet.getInt(13),
-                    )
-                    items.add(item)
-                }
+                extractPostDataClassListFromResultSet(resultSet)
             }
         }
     }
